+++ CHANGELOG.md
@@ -5,15 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.5.0] Add Get-ChocoLogEntry
+
+- Add a new cmdlet that grabs the latest log.
+  [#7](https://github.com/HeyItsGilbert/ChocoLogParse/issues/7)
+  - Has a report flag that may make it convenient for reporting back to tools.
+- Use environment variable `ChocolateyInstall` for default log location.
+  [#8](https://github.com/HeyItsGilbert/ChocoLogParse/issues/8)
+- Replace ArrayList types with GenericList
+  [#9](https://github.com/HeyItsGilbert/ChocoLogParse/issues/9)
+
 ## [0.4.0] Fix external dependency
 
 The dependency on Log4Net was added incorrectly as an external dependency.
 
 ## [0.3.0] Improved Configuration Property
 
-The configuration propertynow supports sub keys to improve filtering.
+The configuration property now supports sub keys to improve filtering.
 
 This also includes a fix to handle multiple threads being mixed.
+[#3](https://github.com/HeyItsGilbert/ChocoLogParse/issues/3)
 
 ## [0.2.0] Formatting